requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
argcomplete==3.1.2
async-timeout==4.0.3
attrs==23.1.0
boto3==1.28.61
botocore==1.31.61
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
contourpy==1.1.0
cycler==0.11.0
dacite==1.7.0
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
fiftyone==0.22.0
fiftyone-brain==0.13.1
fiftyone-db==0.4.0
filelock==3.12.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
glob2==0.7
google-auth==2.22.0
google-auth-oauthlib==1.0.0
graphql-core==3.2.3
grpcio==1.57.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
imageio==2.31.5
inflate64==0.3.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
kaleido==0.2.1
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.9.0
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mongoengine==0.24.2
motor==3.3.1
mpmath==1.3.0
multidict==6.0.4
multivolumefile==0.2.3
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python-headless==4.8.1.78
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
plotly==5.17.0
pprintpp==0.4.0
priority==2.0.0
protobuf==4.24.2
psutil==5.9.5
py7zr==0.20.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycryptodomex==3.19.0
pymongo==4.5.0
pyparsing==3.0.9
pyppmd==1.0.0
python-dateutil==2.8.2
pytorch-lightning==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzstd==0.15.9
rarfile==4.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
s3transfer==0.7.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.31.1
strawberry-graphql==0.138.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
texttable==1.7.0
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.0.1
torchmetrics==1.1.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.1
universal-analytics-python3==1.1.1
urllib3==1.26.16
voxel51-eta==0.12.0
wcwidth==0.2.8
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.9.2
